SignController and UserController Separation
